+++ client/src/components/App.jsx
@@ -41,7 +41,7 @@ class App extends React.Component {
         {page === 'landingPage' && <LandingPage />}
         {page === 'mainPage' && <MainPage user={user} />}
         {page === 'profilePage' && <ProfilePage user={user} />}
-        {page === 'inbox' && <Inbox user={user} />}
+        {page === 'inbox' && <Inbox />}
       </div>
     );
   }
